requirements.txt
# Core dependencies for YOLO training project
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# YOLOv8 (Ultralytics)
ultralytics>=8.0.0

# YOLOv9 dependencies
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=9.5.0
PyYAML>=6.0
requests>=2.31.0
scipy>=1.10.0
seaborn>=0.12.0
pandas>=2.0.0
tqdm>=4.65.0

# Experiment tracking and visualization
wandb>=0.15.0
tensorboard>=2.13.0
tensorboardX>=2.6

# Data processing
albumentations>=1.3.0
imgaug>=0.4.0

# Model evaluation and metrics
scikit-learn>=1.3.0
pycocotools>=2.0.6

# Utilities
click>=8.1.0
rich>=13.0.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# Development tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
pre-commit>=3.3.0

# Optional GPU acceleration
# nvidia-ml-py3>=7.352.0
